Deleted FluxGastrointestinalStromalTumor, FluxBreastCancerDisorderPresent, and FluxSolidTumorCancer classes, moved functions in these classes into FluxCancerDisorderPresent. Edited somes places in the code that was looking for FluxBreastCancerDisorderPresent to use FluxCancerDisorderPresent instead. Still need to add a check for a specific cancer type
