+++ src/model/brca/FluxBrcaObjectFactory.js
@@ -1,7 +1,10 @@
 import { getNamespaceAndName } from '../json-helper';
-import FluxBreastCancerDisorderPresent from './FluxBreastCancerDisorderPresent';
+// import FluxBreastCancerDisorderPresent from './FluxBreastCancerDisorderPresent';
+import FluxCancerDisorderPresent from '../oncocore/FluxCancerDisorderPresent';
 import BrcaObjectFactory from './BrcaObjectFactory';
 
+
+// TODO (nng): Not sure if still need this factory?
 export default class FluxBrcaObjectFactory {
     static createInstance(json, type, patientRecord) {
         const { namespace, elementName } = getNamespaceAndName(json, type);
@@ -10,7 +13,8 @@ export default class FluxBrcaObjectFactory {
         }
         // returns Flux wrapper class if found, otherwise use BrcaObjectFactory
         switch (elementName) {
-            case 'BreastCancerDisorderPresent': return new FluxBreastCancerDisorderPresent(json, patientRecord);
+            // case 'BreastCancerDisorderPresent': return new FluxBreastCancerDisorderPresent(json, patientRecord);
+            case 'BreastCancerDisorderPresent': return new FluxCancerDisorderPresent(json, patientRecord); // nng: not sure about this
             default: return BrcaObjectFactory.createInstance(json, type, patientRecord);
         }
     }
